maint: more renames
