+++ packages/tabarena/src/tabarena/benchmark/exec_models/utils.py
@@ -6,7 +6,6 @@ import numpy as np
 import pandas as pd
 from loguru import logger
 
-from tabarena.benchmark.exec_models.time_split import split_time_index_into_intervals
 from tabarena.benchmark.task.metadata import (
     GroupLabelTypes,
 )
@@ -16,6 +15,119 @@ if TYPE_CHECKING:
         SplitTimeHorizonTypes,
         SplitTimeHorizonUnitTypes,
     )
+
+
+def _make_perm(n: int, seed: int = 0) -> tuple[np.ndarray, np.ndarray]:
+    """Return (perm, inv_perm) for length n, using a deterministic RNG seed."""
+    rng = np.random.default_rng(seed)
+    perm = rng.permutation(n)
+    inv_perm = np.empty_like(perm)
+    inv_perm[perm] = np.arange(n)
+    return perm, inv_perm
+
+
+def _apply_inv_perm(obj, inv_perm: np.ndarray, index: pd.Index | None = None):
+    """Inverse-permute predictions while preserving type (Series/DataFrame/ndarray)."""
+    if isinstance(obj, pd.Series):
+        vals = obj.to_numpy()[inv_perm]
+        return pd.Series(vals, index=index, name=obj.name)
+    if isinstance(obj, pd.DataFrame):
+        vals = obj.to_numpy()[inv_perm, :]
+        return pd.DataFrame(vals, index=index, columns=obj.columns)
+    # Fallback: numpy array or array-like
+    arr = np.asarray(obj)
+    return arr[inv_perm]
+
+
+def split_time_index_into_intervals(
+    *,
+    time_data: pd.Series,
+    goal_n_intervals: int,
+    balance_on: str = "rows",
+) -> tuple[pd.Series, int]:
+    """Split a monotonically ordered time index into contiguous dynamic intervals.
+
+    Rules:
+    - Larger time values are always later in time
+    - Equal time values are always assigned to the same interval
+    - Intervals are created from the observed data, not equal-width spacing
+    - Tries to create `goal_n_intervals`, but falls back to a smaller number if needed
+    - Never returns fewer than 2 intervals
+
+    Parameters
+    ----------
+    time_data : pd.Series
+        Time input
+    goal_n_intervals : int
+        Desired number of intervals.
+    balance_on : {"rows", "unique"}, default "rows"
+        - "rows": balance intervals by number of rows
+        - "unique": balance intervals by number of unique time values
+
+    Returns:
+    -------
+    time_intervals: pd.Series
+        Interval label for each row in the input time_data.
+    actual_n_intervals : int
+        Number of intervals actually used.
+    """
+    if goal_n_intervals < 2:
+        raise ValueError("n_intervals must be at least 2.")
+    if balance_on not in {"rows", "unique"}:
+        raise ValueError("balance_on must be either 'rows' or 'unique'.")
+
+    assert not time_data.isna().any(), "Time column contains nan values!"
+
+    s = time_data.copy()
+
+    # Aggregate identical time values so duplicates stay together
+    counts = s.value_counts(dropna=False).sort_index().rename("row_count").to_frame()
+    counts["unique_weight"] = 1
+
+    n_unique = len(counts)
+    if n_unique < 2:
+        raise ValueError("Need at least 2 unique time values to create at least 2 intervals.")
+    actual_n_intervals = min(goal_n_intervals, n_unique)
+    if actual_n_intervals < 2:
+        raise ValueError("Could not create at least 2 intervals.")
+
+    weight_col = "row_count" if balance_on == "rows" else "unique_weight"
+    weights = counts[weight_col].to_numpy()
+
+    # Greedy partition of sorted unique values into contiguous groups
+    # aiming for equal cumulative weight per interval.
+    total_weight = weights.sum()
+    cut_positions = []
+    start = 0
+    cumulative = np.cumsum(weights)
+
+    for group_num in range(1, actual_n_intervals):
+        target = group_num * total_weight / actual_n_intervals
+
+        # Candidate cut indices are between unique values:
+        # cut after index j means next group starts at j+1
+        min_j = start
+        max_j = n_unique - (actual_n_intervals - group_num) - 1
+
+        # Choose cut whose cumulative weight is closest to target
+        candidates = np.arange(min_j, max_j + 1)
+        j = candidates[np.argmin(np.abs(cumulative[candidates] - target))]
+        cut_positions.append(j)
+        start = j + 1
+
+    # Assign interval labels to each unique time value
+    interval_labels_for_unique = np.empty(n_unique, dtype=int)
+    prev = 0
+    for interval_id, cut in enumerate(cut_positions):
+        interval_labels_for_unique[prev : cut + 1] = interval_id
+        prev = cut + 1
+    interval_labels_for_unique[prev:] = len(cut_positions)
+
+    mapping = pd.Series(interval_labels_for_unique, index=counts.index)
+
+    time_intervals = time_data.map(mapping)
+
+    return time_intervals, actual_n_intervals
 
 
 class TabArenaValidationProtocolExecMixin:
